# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A secure certificate management API for private keys, CSRs, and certificates",
        "title": "Certificate Monkey API",
        "termsOfService": "https://github.com/your-org/certificate-monkey/blob/main/TERMS.md",
        "contact": {
            "name": "Certificate Monkey Support",
            "url": "https://github.com/your-org/certificate-monkey",
            "email": "support@certificate-monkey.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/your-org/certificate-monkey/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/keys": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of certificate entities with optional filtering by tags, status, key type, and date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate Management"
                ],
                "summary": "List certificates with filtering",
                "parameters": [
                    {
                        "enum": [
                            "CSR_CREATED",
                            "CERT_UPLOADED",
                            "EXPIRED",
                            "REVOKED"
                        ],
                        "type": "string",
                        "description": "Filter by certificate status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "RSA2048",
                            "RSA4096",
                            "ECDSA-P256",
                            "ECDSA-P384"
                        ],
                        "type": "string",
                        "description": "Filter by key type",
                        "name": "key_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter certificates created after this date (RFC3339 format)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter certificates created before this date (RFC3339 format)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page (default: 50, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by environment tag",
                        "name": "environment",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by project tag",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by team tag",
                        "name": "team",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of certificate entities",
                        "schema": {
                            "$ref": "#/definitions/models.ListKeysResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a new private key pair and creates a certificate signing request (CSR) with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate Management"
                ],
                "summary": "Create a new private key and certificate signing request",
                "parameters": [
                    {
                        "description": "Certificate creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created private key and CSR",
                        "schema": {
                            "$ref": "#/definitions/models.CreateKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/keys/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific certificate entity including its private key, CSR, and certificate details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate Management"
                ],
                "summary": "Get certificate by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate entity details",
                        "schema": {
                            "$ref": "#/definitions/models.CertificateEntity"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Certificate not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/keys/{id}/certificate": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads and validates a certificate against an existing certificate signing request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificate Management"
                ],
                "summary": "Upload certificate for existing CSR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate entity ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Certificate upload request containing PEM-encoded certificate",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UploadCertificateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/models.UploadCertificateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid certificate or ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing API key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Certificate entity not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateKeyRequest": {
            "type": "object",
            "required": [
                "common_name",
                "key_type"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "common_name": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "email_address": {
                    "type": "string"
                },
                "key_type": {
                    "$ref": "#/definitions/models.KeyType"
                },
                "organization": {
                    "type": "string"
                },
                "organizational_unit": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "subject_alternative_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.ListKeysResponse": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CertificateEntity"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication. Can also be provided as Bearer token in Authorization header.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer token authentication. Format: \"Bearer \u003capi-key\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}